Fix Q_irand() on 64 bit systems
It use to return values outside range min <= value <= max. Fixed many things.

+++ code/game/q_math.c
@@ -1429,7 +1429,7 @@ void NormalToLatLong( const vec3_t normal, byte bytes[2] )
 // eg the typical tint = (rand() * 255) / 32768
 // becomes tint = irand(0, 255)
 
-static unsigned long	holdrand = 0x89abcdef;
+static uint32_t	holdrand = 0x89abcdef;
 
 void Rand_Init(int seed)
 {
